在发送message时附带index，避免了客户端重复收到相同信息的问题

+++ message_factory.h
@@ -9,7 +9,7 @@ class Message_factory{
 public:
     map<pair<int,int>,string*> message_storehouse;//TODO: link sql
     int nowround=0;
-    int nownum=0;
+    long long nownum=0;
     int mod=1e9+7;
     char buff[2048];
